Attempt at fixing command aliases
Currently, due to the way that methods are searched in $pkg:: which
is an unordered hash and how add_handlers() works, behavior of
command aliases is random and incorrect.

(Sometimes aliases are installed, sometimes not, and sometimes they
are categorized as extra/non-command aliases.)

This is happening because add_handlers() is checking for command
aliases before it has seen/processed all the commands.

The suggested fix solves the issue by splitting the loop in two
identical passes, avoiding the issue. Please modify/improve in any
way that would be needed from here.